// src/common/enums/tipo-.conta.enum.ts
export enum TipoConta {
  CORRENTE = 'CORRENTE',
  POUPANCA = 'POUPANCA',
}

export enum TipoCargo {
  AGENTE = 'Agente',
  GERENTE = 'Gerente',
}

export enum TipoTransacao {
  DEBITO = 'debito',
  CREDITO = 'credito',
  DEPOSITO = "DEPOSITO",
  SAQUE = "SAQUE",
  TRANSFERENCIA = "TRANSFERENCIA",
  PAGAMENTO = "PAGAMENTO",
}